requirements.txt
Django>=5.2.8
python-dateutil>=2.9.0
mysqlclient>=2.2.0
# Alternative MySQL driver for Windows (if mysqlclient fails):
# pymysql>=1.1.0
python-dotenv>=1.0.0

